test(unit): move execution mocks to unit conftest
